Moved dtos into api folder
