+++ MyMoney.Web/ClientApp/src/app/shared/api/dtos/budget-list-response.interface.ts
@@ -1,4 +1,4 @@
-import { IBudgetModel } from '../state/types';
+import { IBudgetModel } from '../../state/types';
 
 export interface BudgetListResponse {
    budgets: IBudgetModel[];
