+++ MyMoney.Web/ClientApp/src/app/shared/api/dtos/delete-response.interface.ts
