+++ MyMoney.Web/ClientApp/src/app/shared/api/dtos/id-request.interface.ts
