+++ MyMoney.Web/ClientApp/src/app/shared/api/dtos/login-response.interface.ts
