+++ MyMoney.Web/ClientApp/src/app/shared/api/dtos/register-request.interface.ts
