+++ MyMoney.Web/ClientApp/src/app/shared/api/dtos/transaction-list-response.interface.ts
