+++ MyMoney.Web/ClientApp/src/app/shared/api/dtos/update.response.interface.ts
