+++ MyMoney.Web/ClientApp/src/app/shared/api/index.ts
@@ -1,0 +1,3 @@
+export * from './transaction.api';
+export * from './dtos';
+export * from './budget.api';
