+++ MyMoney.Web/ClientApp/src/app/shared/services/authentication-service.service.ts
@@ -2,11 +2,10 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { first, map } from 'rxjs/operators';
-import { LoginResponse } from '../interfaces/login-response.interface';
+import { LoginResponse, RegisterRequest } from '../api';
 import { Store } from '@ngrx/store';
 import { IAppState } from '../state/app-state';
 import { ClearSessionAction, StartSessionAction } from '../state/actions';
-import { RegisterRequest } from '../interfaces';
 import { selectCurrentSession } from '../state/selectors/session.selector';
 import { ISessionModel } from '../state/types';
 import { Router } from '@angular/router';
