+++ MyMoney.Web/ClientApp/src/app/shared/services/transaction-service.service.ts
@@ -2,8 +2,7 @@ import { Injectable } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Observable, of } from 'rxjs';
 import { concatAll, map } from 'rxjs/operators';
-import { TransactionApi } from '../api/transaction.api';
-import { DeleteResponse, TransactionListResponse, UpdateResponse } from '../interfaces';
+import { TransactionApi, DeleteResponse, TransactionListResponse, UpdateResponse } from '../api';
 import { DeleteTransactionAction, RefreshTransactionsAction, SetTransactionsAction, UpdateDataRangeAction, UpdateTransactionAction } from '../state/actions';
 import { IAppState } from '../state/app-state';
 import { selectTransactionsSearchParameters, selectTransaction } from '../state/selectors/transaction.selector';
